Move index page elemens to `work` folder
